hms/requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.4.0
crashtest==0.4.1
cryptography==41.0.5
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.7
django-allauth==0.58.2
django-browser-reload==1.12.0
django-environ==0.11.2
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.13.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pexpect==4.8.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomlkit==0.12.2
trove-classifiers==2023.11.9
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.6
zipp==3.17.0
